requirements/base.txt
# Python
pillow
uuid

# Django Specific
Django==1.11
django-suit==0.2.23
django-registration-redux==1.4
django-crispy-forms==1.6.1
django-imagekit==3.3
django-extensions==1.6.7
django-toolset==0.1.5
django_compressor==2.1

# AWS & DB
boto==2.45.0
redis==2.10.5
